update: abstractize data-gen
