+++ projects/pswg_gadgets/src/client/java/dev/pswg/datagen/GadgetsDataGenerator.java
@@ -83,12 +83,7 @@ public class GadgetsDataGenerator implements DataGeneratorEntrypoint
 	{
 		public ModelGenerator(FabricDataOutput output)
 		{
-			super(output);
-		}
-
-		public static WeightedVariant createWeightedVariant(Identifier id)
-		{
-			return new WeightedVariant(Pool.of(new ModelVariant(id)));
+			super(output, Gadgets.MODID);
 		}
 
 		@Override
@@ -123,136 +118,10 @@ public class GadgetsDataGenerator implements DataGeneratorEntrypoint
 			}
 		}
 
-		private static void registerAccumulatingBlock(Block block, BlockStateModelGenerator generator)
-		{
-			var id = TextureMap.getId(block);
-
-			generator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateVariantMap.models(Properties.LAYERS).generate(
-					height -> {
-						Identifier modelId = TexturedModel.makeFactory(block1 -> TextureMap.all(id).put(TextureKey.PARTICLE, id), blockModel("template_accumulating_height" + height * 2, TextureKey.ALL, TextureKey.PARTICLE)).upload(block, "_height" + height * 2, generator.modelCollector);
-						return createWeightedVariant(modelId);
-					}))
-			);
-			generator.registerParentedItemModel(block, ModelIds.getBlockSubModelId(block, "_height2"));
-		}
-
-		private static void registerJaporLeaves(Block block, BlockStateModelGenerator generator)
-		{
-			Model model = new Model(Optional.of(Galaxies.id("block/japor_leaves")), Optional.empty(), TextureKey.TEXTURE);
-			TexturedModel texturedModel = TexturedModel.makeFactory(TextureMap::texture, model).get(block);
-			WeightedVariant weightedVariant = createWeightedVariant(texturedModel.upload(block, generator.modelCollector));
-			generator.blockStateCollector.accept(
-					VariantsBlockModelDefinitionCreator.of(block, weightedVariant).apply(BlockStateVariantMap.operations(Properties.HORIZONTAL_FACING)
-					                                                                                         .register(Direction.EAST, BlockStateModelGenerator.ROTATE_Y_90)
-					                                                                                         .register(Direction.SOUTH, BlockStateModelGenerator.ROTATE_Y_180)
-					                                                                                         .register(Direction.WEST, BlockStateModelGenerator.ROTATE_Y_270)
-					                                                                                         .register(Direction.NORTH, BlockStateModelGenerator.NO_OP)));
-			generator.registerItemModel(block);
-		}
-
-		private static void registerLog(Block block, DataGenBlock dataGenBlock, BlockStateModelGenerator generator)
-		{
-			String logKey = getBlockKey(block).toString();
-			BlockStateModelGenerator.LogTexturePool texturePool = generator.createLogTexturePool(block).log(block);
-			if (dataGenBlock.model() == DataGenBlockModel.LogWithWood)
-			{
-				String woodKey = logKey.substring(0, logKey.indexOf("_log")) + "_wood";
-				Block woodBlock = Registries.BLOCK.get(Identifier.of(woodKey));
-				texturePool.wood(woodBlock);
-			}
-		}
-		private static void registerCubeWithRotation(Block block, DataGenBlock dataGenBlock, TexturedModel.Factory modelFactory, BlockStateModelGenerator generator){
-			switch (dataGenBlock.rotation()){
-				case Default -> generator.registerSingleton(block, modelFactory);
-				case RandomRotationX -> {
-					Identifier id = modelFactory.upload(block, generator.modelCollector);
-					var blockStateSupplier = MultipartBlockModelDefinitionCreator.create(block).with(new WeightedVariant(Pool.of(
-							new Weighted<>(new ModelVariant(id).withRotationX(AxisRotation.R0), 1),
-							new Weighted<>(new ModelVariant(id).withRotationX(AxisRotation.R90), 1),
-							new Weighted<>(new ModelVariant(id).withRotationX(AxisRotation.R180), 1),
-							new Weighted<>(new ModelVariant(id).withRotationX(AxisRotation.R270), 1)
-					)));
-					generator.blockStateCollector.accept(blockStateSupplier);
-				}
-				case AxisRotated -> {
-					Identifier id = modelFactory.upload(block, generator.modelCollector);
-					var blockStateSupplier = VariantsBlockModelDefinitionCreator.of(block, createWeightedVariant(id)).apply(BlockStateVariantMap.operations(Properties.AXIS)
-					                                                                                                                            .register(Direction.Axis.Y, ModelVariantOperator.MODEL.withValue(id))
-					                                                                                                                            .register(Direction.Axis.Z, ModelVariantOperator.MODEL.withValue(id)
-					                                                                                                                                                                                  .then(ModelVariantOperator.ROTATION_X.withValue(AxisRotation.R90)))
-					                                                                                                                            .register(Direction.Axis.X, ModelVariantOperator.MODEL.withValue(id)
-					                                                                                                                                                                                  .then(ModelVariantOperator.ROTATION_X.withValue(AxisRotation.R90))
-					                                                                                                                                                                                  .then(ModelVariantOperator.ROTATION_Y.withValue(AxisRotation.R90))
-					                                                                                                                            )
-					);
-					generator.blockStateCollector.accept(blockStateSupplier);
-				}
-			}
-		}
-		private static void registerStairs(Block stairs, BlockStateModelGenerator generator){
-			Identifier stairsId = Models.STAIRS.upload(stairs, TextureMap.all(stairs), generator.modelCollector);
-			Identifier stairsOuterId = Models.OUTER_STAIRS.upload(stairs, TextureMap.all(stairs), generator.modelCollector);
-			Identifier stairsInnerId = Models.INNER_STAIRS.upload(stairs, TextureMap.all(stairs), generator.modelCollector);
-
-			generator.blockStateCollector.accept(BlockStateModelGenerator.createStairsBlockState(stairs, createWeightedVariant(stairsInnerId), createWeightedVariant(stairsId), createWeightedVariant(stairsOuterId)));
-			generator.registerItemModel(stairs);
-
-		}
-
-		private static void registerVerticalSlab(Block slab, BlockStateModelGenerator generator){
-			registerVerticalSlabAllTextures(slab, slab, generator);
-		}
-		private static void registerVerticalSlabAllTextures(Block textureBase, Block slab, BlockStateModelGenerator generator){
-			var textureId = TextureMap.getId(textureBase);
-			var textureMap = new TextureMap().put(TextureKey.SIDE, textureId).put(TextureKey.TOP, textureId).put(TextureKey.END, textureId);
-			Identifier bottomId = Models.SLAB.upload(slab, textureMap, generator.modelCollector);
-			Identifier topId =  Models.SLAB_TOP.upload(slab, "_top", textureMap, generator.modelCollector);
-			Identifier doubleId =  Models.CUBE_COLUMN.upload(slab, "_double", textureMap, generator.modelCollector);
-
-			var blockState = VariantsBlockModelDefinitionCreator.of(slab, createWeightedVariant(bottomId)).
-			                                                    apply(BlockStateVariantMap.operations(Properties.AXIS)
-			                                                                              .register(Direction.Axis.Y, ModelVariantOperator.ROTATION_X.withValue(AxisRotation.R0))
-			                                                                              .register(Direction.Axis.Z, ModelVariantOperator.ROTATION_X.withValue(AxisRotation.R270).then(ModelVariantOperator.UV_LOCK.withValue(true)))
-			                                                                              .register(Direction.Axis.X, ModelVariantOperator.ROTATION_X.withValue(AxisRotation.R90).then(ModelVariantOperator.ROTATION_Y.withValue(AxisRotation.R90)).then(ModelVariantOperator.UV_LOCK.withValue(true)))
-			                                                    ).apply(BlockStateVariantMap.operations(Properties.SLAB_TYPE)
-			                                                                                .register(SlabType.BOTTOM, ModelVariantOperator.MODEL.withValue(bottomId))
-			                                                                                .register(SlabType.DOUBLE, ModelVariantOperator.MODEL.withValue(doubleId))
-			                                                                                .register(SlabType.TOP, ModelVariantOperator.MODEL.withValue(topId))
-					);
-			generator.blockStateCollector.accept(blockState);
-
-		}
-		private static Model blockModel(String parent, TextureKey... requiredTextureKeys)
-		{
-			return new Model(Optional.of(Gadgets.id("block/" + parent)), Optional.empty(), requiredTextureKeys);
-		}
-
-		private static void registerLightingPanel(Block block, BlockStateModelGenerator generator) {
-			Identifier identifier = lightingPanelFactory().upload(block, generator.modelCollector);
-			Identifier identifier2 = generator.createSubModel(block, "_on", Models.CUBE_COLUMN, ModelGenerator::createLightingPanelTextureMap);
-
-			generator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateModelGenerator.createBooleanModelMap(Properties.LIT, createWeightedVariant(identifier2), createWeightedVariant(identifier))));
-		}
-		public static TexturedModel.Factory lightingPanelFactory(){
-			return TexturedModel.makeFactory(ModelGenerator::createLightingPanelTextureMap, Models.CUBE_COLUMN);
-		}
-		public static TextureMap createLightingPanelTextureMap(Block block){
-			return createLightingPanelTextureMap(TextureMap.getId(block));
-		}
-		public static TextureMap createLightingPanelTextureMap(Identifier identifier){
-			return new TextureMap().put(TextureKey.SIDE, identifier).put(TextureKey.END, TextureMap.getId(GalaxiesBlocks.GRAY_IMPERIAL_PANEL_PATTERN_3));
-		}
-
-		public static Identifier getBlockKey(Block block)
-		{
-			return block.getRegistryEntry().getKey().get().getValue();
-		}
-
 		@Override
 		public void generateItemModels(ItemModelGenerator itemModelGenerator)
 		{
 			AutoGenerateUtil.consumeAnnotatedFields(DataGenItem.class, GadgetsItems.class, Item.class, (item, dataGenItem) -> registerItem(itemModelGenerator, item, dataGenItem));
-
 		}
 
 		public static Identifier createItemKey(Item item, DataGenItem dataGenItem)
@@ -275,24 +144,6 @@ public class GadgetsDataGenerator implements DataGeneratorEntrypoint
 						case handheld -> register(generator, item, createItemKey(item, dataGenItem), Models.HANDHELD);
 						case drink -> registerDrink(generator, item, dataGenItem);
 					}
-			}
-		}
-
-		public void registerDrink(ItemModelGenerator generator, Item item, DataGenItem dataGenItem)
-		{
-			Identifier modelId;
-			Identifier overlay = (dataGenItem.overlayTextureOverride().equals("")) ? Identifier.of(createItemKey(item, dataGenItem).withSuffixedPath("_overlay").toString().replace("_filled", "")) : Gadgets.id(dataGenItem.overlayTextureOverride()).withPrefixedPath("item/");
-			Identifier base = Identifier.of(ModelIds.getItemModelId(item).toString().replace("_filled", ""));
-
-			if (dataGenItem.invertLayer())
-			{
-				modelId = generator.uploadTwoLayers(item, base, overlay);
-				generator.output.accept(item, ItemModels.tinted(modelId, new ConstantTintSource(16777215), new SwgDrinkTintSource()));
-			}
-			else
-			{
-				modelId = generator.uploadTwoLayers(item, overlay, base);
-				generator.output.accept(item, ItemModels.tinted(modelId, new SwgDrinkTintSource()));
 			}
 		}
 	}
